refactor: extract PromptSettingsScreen, OAuthLoginSection, and PerformanceStatsDialog
Move private composables from SettingsTab.kt into dedicated files:
- PromptSettingsScreen → ui/screens/PromptSettingsScreen.kt (234 lines)
- OAuthLoginSection + formatExpiration → ui/components/OAuthLoginSection.kt (264 lines)
- PerformanceStatsDialog + formatLastUsed + formatAudioDuration → ui/dialogs/PerformanceStatsDialog.kt (287 lines)

SettingsTab.kt reduced from 2072 to 1395 lines (-33%).
