+++ app/src/main/java/com/antivocale/app/ui/tabs/SettingsTab.kt
@@ -39,10 +39,13 @@ import com.antivocale.app.data.HuggingFaceTokenManager
 import com.antivocale.app.data.HuggingFaceOAuthConfig
 import com.antivocale.app.data.ModelSource
 import com.antivocale.app.ui.components.CollapsibleSection
+import com.antivocale.app.ui.components.OAuthLoginSection
 import com.antivocale.app.ui.components.SettingsDropdown
 import com.antivocale.app.ui.components.ToggleSettingCard
 import com.antivocale.app.ui.components.UnloadModelButton
+import com.antivocale.app.ui.dialogs.PerformanceStatsDialog
 import com.antivocale.app.ui.screens.PerAppSettingsScreen
+import com.antivocale.app.ui.screens.PromptSettingsScreen
 import com.antivocale.app.ui.theme.ThemeType
 import com.antivocale.app.service.InferenceService
 import com.antivocale.app.ui.viewmodel.SettingsViewModel
@@ -1389,684 +1392,4 @@ fun SettingsTab(
         Spacer(modifier = Modifier.height(32.dp))
     }
     } // End of if-else for showPerAppSettings
-}
-
-// ==================== Prompt Settings Screen ====================
-
-@Composable
-private fun PromptSettingsScreen(
-    viewModel: SettingsViewModel,
-    onBack: () -> Unit
-) {
-    val defaultPrompt by viewModel.defaultPrompt.collectAsState()
-    var promptInput by remember { mutableStateOf(defaultPrompt) }
-    val context = LocalContext.current
-
-    LaunchedEffect(defaultPrompt) {
-        promptInput = defaultPrompt
-    }
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .navigationBarsPadding()
-            .verticalScroll(rememberScrollState())
-            .padding(16.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        // Header with back button
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            IconButton(onClick = onBack) {
-                Icon(
-                    imageVector = Icons.Default.ArrowBack,
-                    contentDescription = stringResource(R.string.back)
-                )
-            }
-            Text(
-                text = stringResource(R.string.default_prompt_title),
-                style = MaterialTheme.typography.headlineSmall,
-                fontWeight = FontWeight.Bold
-            )
-        }
-
-        // Description
-        Text(
-            text = stringResource(R.string.default_prompt_description),
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-
-        // Current default info
-        Surface(
-            modifier = Modifier.fillMaxWidth(),
-            color = MaterialTheme.colorScheme.surfaceVariant,
-            shape = MaterialTheme.shapes.small
-        ) {
-            Column(
-                modifier = Modifier.padding(12.dp),
-                verticalArrangement = Arrangement.spacedBy(4.dp)
-            ) {
-                Text(
-                    text = stringResource(R.string.current_default_prompt_label),
-                    style = MaterialTheme.typography.labelMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-                Text(
-                    text = defaultPrompt.ifEmpty { stringResource(R.string.builtin_default_prompt) },
-                    style = MaterialTheme.typography.bodyMedium,
-                    fontStyle = if (defaultPrompt.isEmpty()) androidx.compose.ui.text.font.FontStyle.Italic else androidx.compose.ui.text.font.FontStyle.Normal
-                )
-            }
-        }
-
-        // Compatibility info banner
-        Surface(
-            modifier = Modifier.fillMaxWidth(),
-            color = MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.5f),
-            shape = MaterialTheme.shapes.small
-        ) {
-            Row(
-                modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
-                horizontalArrangement = Arrangement.spacedBy(8.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Info,
-                    contentDescription = null,
-                    modifier = Modifier.size(16.dp),
-                    tint = MaterialTheme.colorScheme.onPrimaryContainer
-                )
-                Text(
-                    text = stringResource(R.string.default_prompt_compatibility_info),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onPrimaryContainer
-                )
-            }
-        }
-
-        // Prompt input field
-        OutlinedTextField(
-            value = promptInput,
-            onValueChange = { newValue ->
-                if (newValue.length <= 500) {
-                    promptInput = newValue
-                }
-            },
-            modifier = Modifier.fillMaxWidth(),
-            placeholder = { Text(stringResource(R.string.default_prompt_placeholder)) },
-            minLines = 3,
-            maxLines = 6,
-            supportingText = {
-                Text(
-                    text = stringResource(R.string.default_prompt_chars, promptInput.length),
-                    modifier = Modifier.fillMaxWidth(),
-                    textAlign = TextAlign.End
-                )
-            },
-            trailingIcon = {
-                if (promptInput != defaultPrompt) {
-                    IconButton(onClick = {
-                        viewModel.saveDefaultPrompt(promptInput)
-                    }) {
-                        Icon(
-                            imageVector = Icons.Default.Check,
-                            contentDescription = stringResource(R.string.save)
-                        )
-                    }
-                }
-            }
-        )
-
-        HorizontalDivider()
-
-        // Example Prompts Section
-        Text(
-            text = stringResource(R.string.example_prompts_title),
-            style = MaterialTheme.typography.titleMedium,
-            fontWeight = FontWeight.Bold
-        )
-        Text(
-            text = stringResource(R.string.example_prompts_description),
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-
-        val examplePrompts = listOf(
-            R.string.example_prompt_transcribe,
-            R.string.example_prompt_summarize,
-            R.string.example_prompt_formal,
-            R.string.example_prompt_translate_en,
-            R.string.example_prompt_translate_it,
-            R.string.example_prompt_notes
-        )
-
-        examplePrompts.forEach { promptResId ->
-            val promptText = stringResource(promptResId)
-            Card(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .clickable {
-                        promptInput = promptText
-                        viewModel.saveDefaultPrompt(promptText)
-                        com.antivocale.app.util.ToastCompat.show(
-                            context,
-                            context.getString(R.string.prompt_applied)
-                        )
-                    },
-                colors = CardDefaults.cardColors(
-                    containerColor = if (defaultPrompt == promptText)
-                        MaterialTheme.colorScheme.primaryContainer
-                    else
-                        MaterialTheme.colorScheme.surfaceVariant
-                )
-            ) {
-                Row(
-                    modifier = Modifier.padding(12.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    if (defaultPrompt == promptText) {
-                        Icon(
-                            imageVector = Icons.Default.CheckCircle,
-                            contentDescription = null,
-                            tint = MaterialTheme.colorScheme.primary,
-                            modifier = Modifier.size(20.dp)
-                        )
-                    }
-                    Text(
-                        text = promptText,
-                        style = MaterialTheme.typography.bodyMedium,
-                        modifier = Modifier.weight(1f)
-                    )
-                }
-            }
-        }
-
-        Spacer(modifier = Modifier.height(32.dp))
-    }
-}
-
-// ==================== OAuth Login Section ====================
-
-/**
- * OAuth login section with login button and status display.
- */
-@Composable
-private fun OAuthLoginSection(
-    oauthState: SettingsViewModel.OAuthState,
-    tokenState: HuggingFaceTokenManager.TokenState,
-    onLoginClick: () -> Unit,
-    onLogoutClick: () -> Unit,
-    onDismissError: () -> Unit
-) {
-    val context = LocalContext.current
-    when (oauthState) {
-        is SettingsViewModel.OAuthState.Idle -> {
-            // Show current token status or login button
-            when (val state = tokenState) {
-                is HuggingFaceTokenManager.TokenState.Valid -> {
-                    var showDetails by remember { mutableStateOf(false) }
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable { showDetails = !showDetails },
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Row(
-                            modifier = Modifier.weight(1f),
-                            verticalAlignment = Alignment.CenterVertically,
-                            horizontalArrangement = Arrangement.spacedBy(8.dp)
-                        ) {
-                            Icon(
-                                imageVector = Icons.Default.CheckCircle,
-                                contentDescription = null,
-                                tint = when (state.authType) {
-                                    HuggingFaceTokenManager.AuthType.OAUTH -> MaterialTheme.colorScheme.tertiary
-                                    HuggingFaceTokenManager.AuthType.MANUAL -> MaterialTheme.colorScheme.primary
-                                },
-                                modifier = Modifier.size(16.dp)
-                            )
-                            Text(
-                                text = when (state.authType) {
-                                    HuggingFaceTokenManager.AuthType.OAUTH -> stringResource(R.string.oauth_connected)
-                                    HuggingFaceTokenManager.AuthType.MANUAL -> stringResource(R.string.token_valid)
-                                },
-                                style = MaterialTheme.typography.labelLarge,
-                                color = when (state.authType) {
-                                    HuggingFaceTokenManager.AuthType.OAUTH -> MaterialTheme.colorScheme.tertiary
-                                    HuggingFaceTokenManager.AuthType.MANUAL -> MaterialTheme.colorScheme.primary
-                                }
-                            )
-                            Icon(
-                                imageVector = if (showDetails) Icons.Default.ExpandLess else Icons.Default.ExpandMore,
-                                contentDescription = if (showDetails) stringResource(R.string.hide_details) else stringResource(R.string.show_details),
-                                tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                                modifier = Modifier.size(18.dp)
-                            )
-                        }
-                        IconButton(onClick = onLogoutClick) {
-                            Icon(
-                                imageVector = Icons.Default.Logout,
-                                contentDescription = stringResource(R.string.logout),
-                                tint = MaterialTheme.colorScheme.error
-                            )
-                        }
-                    }
-                    if (showDetails) {
-                        Column(
-                            modifier = Modifier.padding(start = 24.dp),
-                            verticalArrangement = Arrangement.spacedBy(4.dp)
-                        ) {
-                            Text(
-                                text = stringResource(R.string.username_label, state.username),
-                                style = MaterialTheme.typography.bodyMedium
-                            )
-                            Text(
-                                text = stringResource(R.string.token_label, state.maskedToken),
-                                style = MaterialTheme.typography.bodySmall,
-                                fontFamily = androidx.compose.ui.text.font.FontFamily.Monospace,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant
-                            )
-                            if (state.authType == HuggingFaceTokenManager.AuthType.OAUTH && state.expiresAt != null) {
-                                val expiresText = formatExpiration(state.expiresAt, context)
-                                val isExpiringSoon = state.needsRefresh()
-                                Text(
-                                    text = stringResource(R.string.expires_label, expiresText),
-                                    style = MaterialTheme.typography.bodySmall,
-                                    color = if (isExpiringSoon)
-                                        MaterialTheme.colorScheme.error
-                                    else
-                                        MaterialTheme.colorScheme.onSurfaceVariant
-                                )
-                            }
-                        }
-                    }
-                }
-                else -> {
-                    // No valid token - show login button prominently
-                    Button(
-                        onClick = onLoginClick,
-                        modifier = Modifier.fillMaxWidth(),
-                        colors = ButtonDefaults.buttonColors(
-                            containerColor = MaterialTheme.colorScheme.tertiary
-                        )
-                    ) {
-                        Icon(Icons.Default.Login, contentDescription = null)
-                        Spacer(modifier = Modifier.width(8.dp))
-                        Text(stringResource(R.string.login_with_huggingface))
-                    }
-                    Text(
-                        text = stringResource(R.string.oauth_login_description),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                    // Alternative: Get token manually link
-                    val context = LocalContext.current
-                    TextButton(
-                        onClick = {
-                            val intent = android.content.Intent(
-                                android.content.Intent.ACTION_VIEW,
-                                android.net.Uri.parse(HF_TOKEN_SETTINGS_URL)
-                            )
-                            context.startActivity(intent)
-                        }
-                    ) {
-                        Icon(
-                            imageVector = Icons.AutoMirrored.Filled.OpenInNew,
-                            contentDescription = null,
-                            modifier = Modifier.size(16.dp)
-                        )
-                        Spacer(modifier = Modifier.width(4.dp))
-                        Text(stringResource(R.string.create_account_link))
-                    }
-                }
-            }
-        }
-        is SettingsViewModel.OAuthState.InProgress -> {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.Center,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                CircularProgressIndicator(modifier = Modifier.size(20.dp))
-                Spacer(modifier = Modifier.width(12.dp))
-                Text(stringResource(R.string.authenticating))
-            }
-        }
-        is SettingsViewModel.OAuthState.Success -> {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.Center,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Icon(
-                    Icons.Default.CheckCircle,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.primary,
-                    modifier = Modifier.size(20.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(
-                    text = stringResource(R.string.logged_in_as, oauthState.username),
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.primary
-                )
-            }
-        }
-        is SettingsViewModel.OAuthState.Error -> {
-            Card(
-                colors = CardDefaults.cardColors(
-                    containerColor = MaterialTheme.colorScheme.errorContainer
-                )
-            ) {
-                Row(
-                    modifier = Modifier.padding(12.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Icon(
-                        Icons.Default.Error,
-                        contentDescription = null,
-                        tint = MaterialTheme.colorScheme.error,
-                        modifier = Modifier.size(20.dp)
-                    )
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Text(
-                        text = oauthState.message,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onErrorContainer,
-                        modifier = Modifier.weight(1f)
-                    )
-                    IconButton(onClick = onDismissError) {
-                        Icon(
-                            Icons.Default.Close,
-                            contentDescription = stringResource(R.string.dismiss),
-                            tint = MaterialTheme.colorScheme.onErrorContainer
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * Formats expiration timestamp to a human-readable string.
- */
-private fun formatExpiration(expiresAt: Long, context: Context): String {
-    val now = System.currentTimeMillis()
-    val remaining = expiresAt - now
-
-    return when {
-        remaining <= 0 -> context.getString(R.string.expired)
-        remaining < 60_000 -> "${remaining / 1000}s"
-        remaining < 3_600_000 -> "${remaining / 60_000}m"
-        remaining < 86_400_000 -> "${remaining / 3_600_000}h"
-        else -> "${remaining / 86_400_000}d"
-    }
-}
-
-
-// ==================== Performance Stats Dialog ====================
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun PerformanceStatsDialog(
-    profiles: List<CalibrationProfile>,
-    isTranscribing: Boolean,
-    onDismiss: () -> Unit,
-    onReset: () -> Unit
-) {
-    var showResetConfirm by remember { mutableStateOf(false) }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Speed,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.primary
-                )
-                Text(stringResource(R.string.performance_stats_title))
-            }
-        },
-        text = {
-            if (profiles.isEmpty()) {
-                Text(
-                    text = stringResource(R.string.performance_stats_empty),
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            } else {
-                Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-                    Text(
-                        text = stringResource(R.string.performance_stats_subtitle),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-
-                    // Find slowest speed for relative calculation
-                    val slowestMsPerSec = profiles.maxOf { it.msPerSecondOfAudio }
-                    val fastestProfile = profiles.first()
-
-                    profiles.forEach { profile ->
-                        val isFastest = profile == fastestProfile && profiles.size > 1
-
-                        // Real-time factor: lower ms/s = faster. RTF < 1.0 means faster than real-time.
-                        val rtf = profile.msPerSecondOfAudio / 1000f
-                        val speedLabel = if (rtf <= 1f) {
-                            String.format("%.1fx real-time", 1f / rtf)
-                        } else {
-                            String.format("%.2fx real-time", 1f / rtf)
-                        }
-                        val relativeSpeed = if (slowestMsPerSec > 0 && profiles.size > 1) {
-                            slowestMsPerSec / profile.msPerSecondOfAudio
-                        } else null
-
-                        Card(
-                            modifier = Modifier.fillMaxWidth(),
-                            shape = MaterialTheme.shapes.small,
-                            colors = CardDefaults.cardColors(
-                                containerColor = if (isFastest)
-                                    MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.3f)
-                                else
-                                    MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f)
-                            )
-                        ) {
-                            Column(
-                                modifier = Modifier.padding(10.dp),
-                                verticalArrangement = Arrangement.spacedBy(4.dp)
-                            ) {
-                                // Row 1: Model name + Fastest badge
-                                Row(
-                                    verticalAlignment = Alignment.CenterVertically,
-                                    horizontalArrangement = Arrangement.spacedBy(6.dp)
-                                ) {
-                                    Text(
-                                        text = profile.displayName,
-                                        style = MaterialTheme.typography.titleSmall,
-                                        fontWeight = FontWeight.SemiBold,
-                                        modifier = Modifier.weight(1f),
-                                        maxLines = 1
-                                    )
-                                    if (isFastest) {
-                                        Surface(
-                                            shape = MaterialTheme.shapes.extraSmall,
-                                            color = MaterialTheme.colorScheme.primary
-                                        ) {
-                                            Text(
-                                                text = stringResource(R.string.performance_stats_fastest_badge),
-                                                modifier = Modifier.padding(horizontal = 6.dp, vertical = 2.dp),
-                                                style = MaterialTheme.typography.labelSmall,
-                                                color = MaterialTheme.colorScheme.onPrimary,
-                                                fontWeight = FontWeight.Bold
-                                            )
-                                        }
-                                    }
-                                }
-
-                                // Row 2: Speed + relative
-                                Row(
-                                    verticalAlignment = Alignment.CenterVertically,
-                                    horizontalArrangement = Arrangement.spacedBy(8.dp)
-                                ) {
-                                    Text(
-                                        text = speedLabel,
-                                        style = MaterialTheme.typography.bodyMedium,
-                                        fontWeight = FontWeight.Medium,
-                                        color = if (isFastest) MaterialTheme.colorScheme.primary
-                                        else MaterialTheme.colorScheme.onSurface
-                                    )
-                                    if (relativeSpeed != null) {
-                                        Text(
-                                            text = String.format("(%.1fx)", relativeSpeed),
-                                            style = MaterialTheme.typography.labelSmall,
-                                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                                        )
-                                    }
-                                }
-
-                                // Row 3: Metadata
-                                Row(
-                                    horizontalArrangement = Arrangement.spacedBy(12.dp)
-                                ) {
-                                    Text(
-                                        text = stringResource(
-                                            R.string.performance_stats_samples_count,
-                                            profile.sampleCount
-                                        ),
-                                        style = MaterialTheme.typography.labelSmall,
-                                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                                    )
-                                    if (profile.lastTimestamp > 0) {
-                                        Text(
-                                            text = formatLastUsed(profile.lastTimestamp),
-                                            style = MaterialTheme.typography.labelSmall,
-                                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                                        )
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    // Total audio processed
-                    val totalAudio = profiles.sumOf { it.totalAudioSeconds }
-                    if (totalAudio > 0) {
-                        HorizontalDivider(modifier = Modifier.padding(vertical = 2.dp))
-                        Text(
-                            text = stringResource(
-                                R.string.performance_stats_total_audio,
-                                formatAudioDuration(totalAudio)
-                            ),
-                            style = MaterialTheme.typography.labelSmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                if (profiles.isNotEmpty()) {
-                    TextButton(onClick = { showResetConfirm = true }) {
-                        Text(
-                            text = stringResource(R.string.performance_stats_clear),
-                            color = MaterialTheme.colorScheme.error
-                        )
-                    }
-                }
-                TextButton(onClick = onDismiss) {
-                    Text(stringResource(R.string.dismiss))
-                }
-            }
-        }
-    )
-
-    // Reset confirmation dialog
-    if (showResetConfirm) {
-        AlertDialog(
-            onDismissRequest = { showResetConfirm = false },
-            title = { Text(stringResource(R.string.performance_stats_clear)) },
-            text = {
-                if (isTranscribing) {
-                    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                        Text(stringResource(R.string.performance_stats_clear_confirm))
-                        HorizontalDivider()
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            horizontalArrangement = Arrangement.spacedBy(8.dp)
-                        ) {
-                            Icon(
-                                imageVector = Icons.Default.Warning,
-                                contentDescription = null,
-                                tint = MaterialTheme.colorScheme.error,
-                                modifier = Modifier.size(18.dp)
-                            )
-                            Text(
-                                text = stringResource(R.string.warn_reset_stats_during_transcription),
-                                style = MaterialTheme.typography.bodySmall,
-                                color = MaterialTheme.colorScheme.error
-                            )
-                        }
-                    }
-                } else {
-                    Text(stringResource(R.string.performance_stats_clear_confirm))
-                }
-            },
-            confirmButton = {
-                TextButton(
-                    onClick = {
-                        onReset()
-                        showResetConfirm = false
-                    }
-                ) {
-                    Text(
-                        text = stringResource(R.string.performance_stats_clear),
-                        color = MaterialTheme.colorScheme.error
-                    )
-                }
-            },
-            dismissButton = {
-                TextButton(onClick = { showResetConfirm = false }) {
-                    Text(stringResource(R.string.action_cancel))
-                }
-            }
-        )
-    }
-}
-
-@Composable
-private fun formatLastUsed(timestamp: Long): String {
-    if (timestamp <= 0) return ""
-    val elapsed = System.currentTimeMillis() - timestamp
-    val minutes = elapsed / 60_000
-    val hours = elapsed / 3_600_000
-    val days = elapsed / 86_400_000
-
-    return when {
-        minutes < 1 -> stringResource(R.string.performance_stats_just_now)
-        minutes < 60 -> stringResource(R.string.performance_stats_minutes_ago, minutes)
-        hours < 24 -> stringResource(R.string.performance_stats_hours_ago, hours)
-        else -> stringResource(R.string.performance_stats_days_ago, days)
-    }
-}
-
-private fun formatAudioDuration(totalSeconds: Long): String {
-    val hours = totalSeconds / 3600
-    val minutes = (totalSeconds % 3600) / 60
-    val seconds = totalSeconds % 60
-
-    return when {
-        hours > 0 -> String.format("%dh %dm", hours, minutes)
-        minutes > 0 -> String.format("%dm %ds", minutes, seconds)
-        else -> String.format("%ds", seconds)
-    }
 }
